Config in common package is specific to repo-fetcher
